feat: make FacilitiesIcons (from PlaceInfo) into common component
