+++ components/common/PlaceInfo.tsx
@@ -1,17 +1,11 @@
-import Image from 'next/image'
 import Link from 'next/link'
 import styled from '@emotion/styled'
 import { PlaceInfoType } from '../../types'
 import { conceptColors } from './BaseLayout'
 import Text from './Text'
+import FacilitiesIcons from './FacilitiesIcons'
 
 const PlaceInfo = ({ place }: { place: PlaceInfoType }) => {
-  const IconImage = ({ src }: { src: string }) => (
-    <IconWrapper>
-      <Image src={src} width={30} height={30} />
-    </IconWrapper>
-  )
-
   return (
     <Link href={`detail/${place._id}`}>
       <PlaceInfoContainer>
@@ -29,12 +23,7 @@ const PlaceInfo = ({ place }: { place: PlaceInfoType }) => {
         <Text color="#b0b0b0" size={0.8}>
           {place.address}
         </Text>
-        <IconsSection>
-          {place.isToiletExists && <IconImage src="/images/toilet.svg" />}
-          {place.isChargerExists && <IconImage src="/images/charger.svg" />}
-          {place.isElevatorExists && <IconImage src="/images/elevator.svg" />}
-          {place.isSlopeExists && <IconImage src="/images/slope.svg" />}
-        </IconsSection>
+        <FacilitiesIcons place={place} />
       </PlaceInfoContainer>
     </Link>
   )
@@ -58,14 +47,6 @@ const NameTypeSection = styled.section`
   div:last-child {
     margin-left: 0.4rem;
   }
-`
-const IconsSection = styled.section`
-  display: flex;
-  margin-top: 0.5rem;
-  justify-content: flex-end;
-`
-const IconWrapper = styled.div`
-  margin: 0 0.1rem;
 `
 
 export default PlaceInfo
